requirements.txt
--index-url https://download.pytorch.org/whl/cpu
aiohttp==3.10.10
fastapi==0.115.3
groq==0.11.0
moviepy==1.0.3
numpy==1.26.3
opencv-python-headless==4.10.0.84
pinecone==5.3.1
pinecone-client==5.0.1
pydub==0.25.1
scipy==1.14.1
soundfile==0.12.1
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
transformers==4.46.0
uvicorn==0.32.0
faiss-cpu==1.9.0
